[📦 refactor] 컴포넌트, staletime 수정 - 페이지네이션 컴포넌트 적용 - staletime 적용
