+++ app/routes/admin/points/readPage.tsx
@@ -1,7 +1,7 @@
 import ReadComponent from "~/components/admin/points/readComponent";
 import {useParams} from "react-router";
 import {useQuery} from "@tanstack/react-query";
-import {readCoupon} from "~/api/points/pointsAPI";
+import {readCoupon} from "~/api/points/adminPointsAPI";
 
 
 function ReadPage() {
